# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dreamsdesign Sales War Machine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /icp:
    get:
      operationId: listIcps
      tags: [icp]
      summary: List all ICPs
      responses:
        "200":
          description: List of ICPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Icp"
    post:
      operationId: createIcp
      tags: [icp]
      summary: Create ICP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIcpBody"
      responses:
        "201":
          description: Created ICP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icp"
  /icp/generate:
    post:
      operationId: generateIcpSuggestions
      tags: [icp]
      summary: Generate ICP suggestions using AI based on website and product description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateIcpBody"
      responses:
        "200":
          description: Array of generated ICP suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IcpSuggestion"
  /icp/{id}:
    get:
      operationId: getIcp
      tags: [icp]
      summary: Get ICP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ICP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icp"
        "404":
          description: Not found
    patch:
      operationId: updateIcp
      tags: [icp]
      summary: Update ICP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIcpBody"
      responses:
        "200":
          description: Updated ICP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icp"
        "404":
          description: Not found
    delete:
      operationId: deleteIcp
      tags: [icp]
      summary: Delete ICP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─────────────────────────────────────────
  # LEADS
  # ─────────────────────────────────────────
...
components:
  schemas:
...
    Icp:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        markets:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        companySize:
          type: string
        filters:
          type: object
        active:
          type: boolean
        leadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, markets, industries, roles, companySize, filters, active, leadCount, createdAt]
    GenerateIcpBody:
      type: object
      properties:
        website:
          type: string
        description:
          type: string
      required: [description]
    IcpSuggestion:
      type: object
      properties:
        name:
          type: string
        markets:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        companySize:
          type: string
        filters:
          type: object
      required: [name, markets, industries, roles, companySize, filters]
    CreateIcpBody:
      type: object
      properties:
        name:
          type: string
        markets:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        companySize:
          type: string
        filters:
          type: object
        active:
          type: boolean
      required: [name, markets, industries, roles, companySize, filters]
    UpdateIcpBody:
      type: object
      properties:
        name:
          type: string
        markets:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        companySize:
          type: string
        filters:
          type: object
        active:
          type: boolean

    # ─── LEADS ───
...